Add simulation updates and bins for creating charts.
